requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.7.0
faiss-cpu==1.13.2
numpy
scikit-learn
pydantic==1.10.13
